Improve chat page and fix mobile Spongie mark
Removed the background and replaced it with a full screen chat
window that should be nicer to use.

There is a weird bug that is happening only **sometimes** on Firefox:
It will show two scroll bars, one for the footer (intended) and
one for the iframe content, even though it fully fits in the iframe.
There seems to be no way to remove it and it doesn't always happen.

I haven't found a way to reproduce it reliably and it works as
expected on all other browsers I've tested.