Use Eyeglass for importing Bootstrap and Bootswatch
The existing styles need to be reviewed, and made to fit in now that we have access
to Bootstrap variables.

SpongeDownloads is a good reference point for this work, and ultimately SpongeHome
will be upgraded to Bootstrap v4 - at which point the stylesheets will be spun off
so they can be shared across the Sponge web properties (home, downloads, and ore).